[Silver III] Title: 01타일, Time: 4 ms, Memory: 5928 KB -BaekjoonHub

+++ 백준/Silver/1904. 01타일/01타일.cc
@@ -12,7 +12,7 @@
 
 using namespace std;
 
-unsigned long long dp[1000001];
+int dp[1000001];
 
 void solve(int n) {
     dp[1] = 1;
